Fixed a bug with time system and changed the data.timeCost to int
after I change the data.timeCost in promptHandler.cs to int, the prompt is not working properly, this need to be fixed

+++ Doctor Game/Assets/Scripts/PromptHandler.cs
@@ -40,7 +40,7 @@ public class PromptHandler : MonoBehaviour
         prompt.transform.GetChild(4).GetComponent<Text>().text = "Time: " + timeCost;
         prompt.transform.GetChild(7).GetComponent<Text>().text = "Stress: " + stressBuff;
 
-        data = new PromptData(float.Parse(staminaCost.Substring(1)), float.Parse(stressCost.Substring(1)), float.Parse(timeCost.Substring(1, timeCost.IndexOf('h') - 1)));
+        data = new PromptData(float.Parse(staminaCost.Substring(1)), float.Parse(stressCost.Substring(1)), int.Parse(timeCost.Substring(1, timeCost.IndexOf('h') - 1)));
 
         prompt.gameObject.SetActive(true);
         targetScene = scene;
@@ -64,9 +64,9 @@ struct PromptData
 {
     public float stressCost;
     public float staminaCost;
-    public float timeCost;
+    public int timeCost;
 
-    public PromptData(float staC, float strC, float tC)
+    public PromptData(float staC, float strC, int tC)
     {
         stressCost = strC;
         staminaCost = staC;
